fix(header): правки после ревью
Jira issue: VEGA-567
