+++ postcss.config.js
@@ -1,7 +1,6 @@
 const createPostcssConfig = require('@gpn-prototypes/frontend-configs/postcss.config');
 
 module.exports = {
-  // eslint-disable-next-line global-require
-  ...require('@gpn-prototypes/frontend-configs/postcss.config'),
+  ...createPostcssConfig(),
   plugins: [...createPostcssConfig().plugins],
 };
