+++ src/components/Header/types.ts
@@ -1,0 +1,9 @@
+import { NavItemType } from '../BaseHeader';
+
+export interface NavLinkType extends NavItemType {
+  url?: string;
+}
+
+export type Params = {
+  projectId?: string;
+};
